requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
APScheduler==3.10.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.7
decorator==5.1.1
docopt==0.6.2
exceptiongroup==1.1.1
executing==1.2.0
fastjsonschema==2.18.0
frozenlist==1.4.0
h11==0.14.0
idna==3.4
ipython==8.13.2
jedi==0.18.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
lxml==4.9.3
matplotlib-inline==0.1.6
multidict==6.0.4
nbformat==5.9.2
numpy==1.24.3
outcome==1.2.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pigar==2.1.1
pipreqs==0.4.13
platformdirs==3.10.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.2
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
wcwidth==0.2.6
webdriver-manager==3.8.6
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2
